trig_array: restore previous behaviour of operator==, optimize a bit operator> and drop unused operator>.
We want the previous behaviour for operator== because when adding arguments to the series
there is not really a way to do it safely by acting on existing terms. We were using the sorted
index to iterate, but forgot the detail that adding argument in front of existing arguments
will change the order in case of equal coefficients.

The only safe way seems to me to create a new series and swap().


git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@34 f32a13d4-301a-0410-ba96-7b22a51b2342