New paradigm for handling series with different arguments.
Now we have append* and prepend* methods which take as arguments vectors of pointers.
to psymbols. Besided, when doing such operations we do not modify() anymore, instead
create tmp series and swap at the end. This should be far less problematic, because
with modify() we were on thin ice regarding which index to use.

Only np work at the moment.


git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@36 f32a13d4-301a-0410-ba96-7b22a51b2342