Fix bug in base_pseries::insert when dealing with non-numerical cf.
If the cf has non-zero arguments we must make sure it is the right size before
checking if it is zero, since we are passing cf_s_vec as argument to the
is_zero function. If there is a mismatch between cf_s_vec size and coefficient's
size, cf's evaluation explodes.


git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@139 f32a13d4-301a-0410-ba96-7b22a51b2342