Accept wider monomials when inserting into polynomial.
The rationale behind this, which is not allowed for terms in pseries,
is that we do not care about width until we insert into series, where we
have the constraints given by argument vectors. All the checks should go
there and polynomial should be free to do as it pleases. This resolves also
a lot of issues e.g. in math::besselJ, where retval is initialized to 1
(hence zero width) and then it is multiplied/summed with non-zero polynomials.


git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@141 f32a13d4-301a-0410-ba96-7b22a51b2342