Modified base_pseries::swap to accept Derived type.
This is necessary because when used from ps it requested base_pseries
type to work, and this made the method useless in Pyranha and possibly
created other problems too with hooks.

Now we must decide if we should never ever deal with base_pseries as
arguments in base_pseries methods. Maybe this is necessary only for
methods called directly from ps (such as swap()). After all we are never
going to use base_pseries directly, hence base_pseries as a type is quite
useless.

Mhrm.


git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@206 f32a13d4-301a-0410-ba96-7b22a51b2342