Enable fast evaluation of trigonometric parts.
Powers of complex exponentials of trigonometric arguments evaluated at t are
cached in a trig_evaluator object and multiplied to provide the complex exponential
of the linear combination of trigonometric arguments in each term. This leads to an
evaluation speedup which can amount up to x2-x3.

Old evaluation routine is still available as base_pseries::t_eval_brute.


git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@410 f32a13d4-301a-0410-ba96-7b22a51b2342