Check for viabiliy of gl_rep on the coding vector elements, not on the codes themselves.
The problem was that we could have ok codes but out-of-range coding vector. There's no really
way around, we must sacrifice half intger range to do this. I've tried with long long int but it
is just dog slow, at that point it is better not to use coded representation.