Fat ugly commit. See below.
Replace math operators in double_cf with explicit methods (mult_by_int(), divide_by_int(), etc.). I
did this because operators can be less flexible (for instance we may need to pass parameters to
operator*= when multiplying polynomials for truncation) and also they are messy when inheriting.

The only surviving operator should be operator=. Not sure about it.

Also, replace *=-1 with invert_sign().