Simplify term insertion.
Now there is one single insertion function, which is templated against Cf2 and two bools (Sign and
CheckTrigSign). A helper struct ensures we are not copying around terms when Cf2 == cf_type. A
wrapper provides a Cf2-only-templated function with positive insertion and trigsigncheck enabled.

More wrappers and more templated bool flags are possible for the future.

When we get around re-fixing pyranha after recent changes other parts of the library will probably
have to be adapted as well.