Introduce int_array class.
This class is a like std::valarray<int>, with these differences:
- lets you choose the bit size of the integer (16 or 8, at the moment)
- lets you choose the signedness of the integer
- offers a hasher function
- implements packed integer operations (on hasher, operator==, initialization and copy)

By sacrificing the max size of the array, which becomes ~32000 instead of ~4E9, we can store an
additional member with no additional weight wrt to std::valarray<int> (i.e., 8 bytes on 32-bits
archs). This additional member caches the "packed size". The unpacked members are reached through
use of bitshift operations, so it should be quite fast.

The bottom line is that some speed-critical operations (hasher, comparison) can be performed in
groups of integers instead of integer-per-integer. I.e, 2 16-bit integers at a time on 32-bits
arch, or 8 8-bits integer at a time on 64-bits archs.