Another - better - try at pow toolbox.
The idea is now to have a named and an anonymous API, so that we can write toolboxes that can be used both in
for coefficient series and top-level series. The anonymous API is defined in base_series and hence can be used
both from cf_series and named_series. For those methods for which name clashes are possible (eval, norm, ...),
the anonymous API should be indicated by the prefix "a_":

base_series::a_eval(const double &, const ArgsTuple);
named_serie::eval(const double &);