Start thinking about a better implementation for proxies.
Basically now we have a proxy term class which corresponds to the real term class
rebound to the proxy types for coefficient and key. Default implementations for such
proxies simply re-declare the typs themselves, so that the proxy term is a 1:1 mapping
to the real term. This allows, for now, to ditch the extra caching of coefficients
in multipliers.