Back to prime sizes and bucket size 6.
There's nothing to do apparently, this time the problem arose in binomial2000,
where we got max load factors of 1E-3 (!) on resize because of some kind of
commensurability in the Kcodes wrt power-of-two sizes.

The moral of the story seems to be that there is a good reason if most hash table
implementations use prime sizes.